requirements.txt
asgiref==3.8.1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
firebase==4.0.1
firebase-admin==6.6.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
proto-plus==1.25.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3

asgiref==3.8.1
astroid==3.3.6
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
dill==0.3.9
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.8
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
idna==3.10
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pylint==3.3.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
